requirements.txt
flask
flask-cors
sqlalchemy
psycopg2-binary
requests
python-dotenv
pyjwt
pytest
